requirements.txt
amqp==2.6.1
asgiref==3.2.10
attrs==19.3.0
billiard==3.6.3.0
celery==4.4.7
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==5.2.1
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.0
django-cors-headers==3.4.0
django-debug-toolbar==2.2
django-extensions==3.0.5
django-filter==2.3.0
django-import-export==2.3.0
django-rest-swagger==2.2.0
djangorestframework==3.11.1
djangorestframework-camel-case==1.2.0
et-xmlfile==1.0.1
Faker==0.9.1
graphviz==0.14.1
idna==2.10
importlib-metadata==1.7.0
iniconfig==1.0.1
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.2
kombu==4.6.11
MarkupPy==1.14
MarkupSafe==1.1.1
mixer==6.1.3
more-itertools==8.4.0
numpy==1.19.1
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.0.4
packaging==20.4
pandas==1.1.0
Pillow==7.2.0
pluggy==0.13.1
psycopg2-binary==2.8.5
py==1.9.0
pyparsing==2.4.7
pytest==6.0.1
pytest-cov==2.10.1
pytest-django==3.9.0
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
redis==3.5.3
requests==2.24.0
sentry-sdk==0.16.5
simplejson==3.17.2
six==1.15.0
sqlparse==0.3.1
tablib==2.0.0
text-unidecode==1.2
toml==0.10.1
uritemplate==3.0.1
urllib3==1.25.10
vine==1.3.0
xlrd==1.2.0
xlwt==1.3.0
zipp==3.1.0
